CI: Actually enable the default fallback behavior of docker-pushrm
The Github action explicitly sets `README.md` as the default file to
push, see
https://github.com/christian-korneck/update-container-description-action/blob/master/action.yml#L17

This disables the fallback to `README-containers.md`, as implemented
in the actual tool that the Github action uses, i.e.
https://github.com/christian-korneck/docker-pushrm

However, by setting the file name explicitly to an empty string, we
can trigger the default fallback behavior of dockre-pushrm after all.

Signed-off-by: beorn7 <beorn@grafana.com>